refactor Auth component and add TabNavigation and TabView for improved dashboard layout; restructure sign-in and check-in utilities
